resolve cyclic deps
