+++ lib/docker-utils.js
@@ -3,18 +3,20 @@
 const isEmpty = require('101/is-empty')
 const WorkerStopError = require('error-cat/errors/worker-stop-error')
 const log = require('./logger')()
+const Docker = require('./docker')
 const Swarm = require('./swarm')
+const rabbitmq = require('./rabbitmq')
 
-/**
- * Do docker top on a random container. All errors would be catched and logged
- * without propagation
- * This should emit the `top` event.
- * We ignore errors since we will restart if event failed
- * @param  {Object} dockerClient instance of Docker or Swarm client (loki)
- * @returns promise
- * @resolves (null) when test complete
- */
 module.exports = class Utils {
+  /**
+   * Do docker top on a random container. All errors would be catched and logged
+   * without propagation
+   * This should emit the `top` event.
+   * We ignore errors since we will restart if event failed
+   * @param  {Object} dockerClient instance of Docker or Swarm client (loki)
+   * @returns promise
+   * @resolves (null) when test complete
+   */
   static testEvent (dockerClient) {
     log.info('test-event')
     return dockerClient.listContainersAsync({
@@ -38,15 +40,6 @@ module.exports = class Utils {
     })
   }
 
-  static toDockerHost (url) {
-    return url.replace('http://', '')
-  }
-
-  static toDockerUrl (host) {
-    const ensuredHost = Utils.toDockerHost(host)
-    return 'http://' + ensuredHost
-  }
-
   /**
    * handles inspect errors.
    * if 404 call task fatal
@@ -61,7 +54,7 @@ module.exports = class Utils {
    *         {Error}           If error unknown
    */
   static _handleInspectError (host, err, log) {
-    const dockerHost = Utils.toDockerHost(host)
+    const dockerHost = Docker.toDockerHost(host)
     if (err.statusCode === 404) {
       // container is not there anymore. Exit
       const fatalErr = new WorkerStopError(
@@ -82,10 +75,8 @@ module.exports = class Utils {
       })
       .then((exists) => {
         if (!exists) {
-          // TODO: cyclic dependency
-          const rabbitmq = require('./rabbitmq')
           rabbitmq.publishEvent('dock.lost', {
-            host: Utils.toDockerUrl(dockerHost)
+            host: Docker.toDockerUrl(dockerHost)
           })
           log.trace('_handleInspectError - host does not exist')
           const fatalErr = new WorkerStopError(
