+++ lib/docker.js
@@ -2,8 +2,6 @@
 require('loadenv')()
 
 const DockerClient = require('@runnable/loki').Docker
-
-const dockerUtils = require('./docker-utils')
 const logger = require('./logger')()
 
 module.exports = class Docker extends DockerClient {
@@ -13,7 +11,16 @@ module.exports = class Docker extends DockerClient {
    * @return {Docker}      Docker instance
    */
   constructor (host) {
-    const dockerHost = 'https://' + dockerUtils.toDockerHost(host)
+    const dockerHost = 'https://' + Docker.toDockerHost(host)
     super({ host: dockerHost, log: logger })
+  }
+
+  static toDockerHost (url) {
+    return url.replace('http://', '')
+  }
+
+  static toDockerUrl (host) {
+    const ensuredHost = Docker.toDockerHost(host)
+    return 'http://' + ensuredHost
   }
 }
