+++ lib/listener.js
@@ -147,7 +147,7 @@ module.exports = class EventListener {
     event.dockerPort = event.Host.split(':')[1]
     // keep tags for legacy reasons
     event.tags = event.org
-    const dockerUrl = dockerUtils.toDockerUrl(event.Host)
+    const dockerUrl = Docker.toDockerUrl(event.Host)
 
     // we expose one value as two results for compatibility reasons
     event.host = dockerUrl
