+++ lib/rabbitmq.js
@@ -2,7 +2,7 @@
 require('loadenv')()
 
 const RabbitMQ = require('ponos/lib/rabbitmq')
-const dockerUtils = require('./docker-utils')
+const Docker = require('./docker')
 const logger = require('./logger')()
 
 class Publisher extends RabbitMQ {
@@ -35,7 +35,7 @@ class Publisher extends RabbitMQ {
     log.info('call')
     // we need to send tags for backwards compatibility
     this.publishEvent(type + '.events-stream.connected', {
-      host: dockerUtils.toDockerUrl(host),
+      host: Docker.toDockerUrl(host),
       org: org,
       tags: org
     })
@@ -55,7 +55,7 @@ class Publisher extends RabbitMQ {
     log.info('call')
     // we need to send tags for backwards compatibility
     this.publishEvent('docker.events-stream.disconnected', {
-      host: dockerUtils.toDockerUrl(host),
+      host: Docker.toDockerUrl(host),
       org: org
     })
   }
